init: 폴더 구조 변경
- 폴더 구조 변경
- repository 추가
