+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/request/SeatUpdateRequest.java
@@ -1,0 +1,11 @@
+package com.spoticket.teamstadium.application.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record SeatUpdateRequest(
+    @NotNull String section,
+    @NotNull Long quantity,
+    @NotNull Integer price
+) {
+
+}
