+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/request/TeamCreateRequest.java
@@ -1,0 +1,17 @@
+package com.spoticket.teamstadium.application.dto.request;
+
+import com.spoticket.teamstadium.domain.model.TeamCategoryEnum;
+import jakarta.validation.constraints.NotNull;
+
+public record TeamCreateRequest(
+
+    @NotNull String name,
+    @NotNull TeamCategoryEnum category,
+    @NotNull String description,
+    String profile,
+    String homeLink,
+    String snsLink
+
+) {
+
+}
