+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/request/TeamUpdateRequest.java
@@ -1,0 +1,14 @@
+package com.spoticket.teamstadium.application.dto.request;
+
+import jakarta.validation.constraints.NotNull;
+
+public record TeamUpdateRequest(
+
+    @NotNull String name,
+    @NotNull String description,
+    String profile,
+    String homeLink,
+    String snsLink
+) {
+
+}
