+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/response/GameReadResponse.java
@@ -1,0 +1,16 @@
+package com.spoticket.teamstadium.application.dto.response;
+
+import com.spoticket.teamstadium.domain.model.TeamCategoryEnum;
+import java.time.LocalDateTime;
+import java.util.UUID;
+
+public record GameReadResponse(
+    UUID gameId,
+    String title,
+    LocalDateTime startTime,
+    TeamCategoryEnum sport,
+    String league
+
+) {
+
+}
