+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/response/StadiumInfoResponse.java
@@ -1,0 +1,23 @@
+package com.spoticket.teamstadium.application.dto.response;
+
+import com.spoticket.teamstadium.domain.model.Stadium;
+import java.util.UUID;
+
+public record StadiumInfoResponse(
+    UUID stadiumId,
+    String name,
+    String address,
+    String seatImage,
+    String description
+) {
+
+  public static StadiumInfoResponse from(Stadium stadium) {
+    return new StadiumInfoResponse(
+        stadium.getStadiumId(),
+        stadium.getName(),
+        stadium.getAddress(),
+        stadium.getSeatImage(),
+        stadium.getDescription()
+    );
+  }
+}
