+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/response/StadiumReadResponse.java
@@ -1,0 +1,10 @@
+package com.spoticket.teamstadium.application.dto.response;
+
+import java.util.List;
+
+public record StadiumReadResponse(
+    StadiumInfoResponse stadium,
+    List<GameReadResponse> games
+) {
+
+}
