+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/application/dto/response/TeamReadResponse.java
@@ -1,0 +1,10 @@
+package com.spoticket.teamstadium.application.dto.response;
+
+import java.util.List;
+
+public record TeamReadResponse(
+    TeamInfoResponse team,
+    List<GameReadResponse> games
+) {
+
+}
