--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/common/ApiResponse.java
@@ -1,12 +1,0 @@
-package com.spoticket.teamstadium.common;
-
-import com.fasterxml.jackson.annotation.JsonInclude;
-
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public record ApiResponse<T>(
-    int status,
-    String msg,
-    T data
-) {
-
-}
