--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/domain/TeamCategoryEnum.java
@@ -1,14 +1,0 @@
-package com.spoticket.teamstadium.domain;
-
-public enum TeamCategoryEnum {
-
-  HANDBALL,
-  ICE_HOCKEY,
-  SOCCER,
-  BASEBALL,
-  BASKETBALL,
-  VOLLEYBALL,
-  OTHER,
-  E_SPORTS
-
-}
