+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/domain/model/TeamCategoryEnum.java
@@ -1,0 +1,14 @@
+package com.spoticket.teamstadium.domain.model;
+
+public enum TeamCategoryEnum {
+
+  HANDBALL,
+  ICE_HOCKEY,
+  SOCCER,
+  BASEBALL,
+  BASKETBALL,
+  VOLLEYBALL,
+  OTHER,
+  E_SPORTS
+
+}
