+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/domain/repository/FavTeamRepository.java
@@ -1,0 +1,5 @@
+package com.spoticket.teamstadium.domain.repository;
+
+public interface FavTeamRepository {
+
+}
