+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/domain/repository/TeamRepository.java
@@ -1,0 +1,11 @@
+package com.spoticket.teamstadium.domain.repository;
+
+import com.spoticket.teamstadium.domain.model.Team;
+import java.util.Optional;
+
+public interface TeamRepository {
+
+  Team save(Team team);
+
+  Optional<Team> findByName(String name);
+}
