--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/dto/request/SeatUpdateRequest.java
@@ -1,11 +1,0 @@
-package com.spoticket.teamstadium.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public record SeatUpdateRequest(
-    @NotNull String section,
-    @NotNull Long quantity,
-    @NotNull Integer price
-) {
-
-}
