--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/dto/request/TeamCreateRequest.java
@@ -1,17 +1,0 @@
-package com.spoticket.teamstadium.dto.request;
-
-import com.spoticket.teamstadium.domain.TeamCategoryEnum;
-import jakarta.validation.constraints.NotNull;
-
-public record TeamCreateRequest(
-
-    @NotNull String name,
-    @NotNull TeamCategoryEnum category,
-    @NotNull String description,
-    String profile,
-    String homeLink,
-    String snsLink
-
-) {
-
-}
