--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/dto/response/GameReadResponse.java
@@ -1,16 +1,0 @@
-package com.spoticket.teamstadium.dto.response;
-
-import com.spoticket.teamstadium.domain.TeamCategoryEnum;
-import java.time.LocalDateTime;
-import java.util.UUID;
-
-public record GameReadResponse(
-    UUID gameId,
-    String title,
-    LocalDateTime startTime,
-    TeamCategoryEnum sport,
-    String league
-
-) {
-
-}
