--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/dto/response/StadiumReadResponse.java
@@ -1,10 +1,0 @@
-package com.spoticket.teamstadium.dto.response;
-
-import java.util.List;
-
-public record StadiumReadResponse(
-    StadiumInfoResponse stadium,
-    List<GameReadResponse> games
-) {
-
-}
