--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/dto/response/TeamReadResponse.java
@@ -1,10 +1,0 @@
-package com.spoticket.teamstadium.dto.response;
-
-import java.util.List;
-
-public record TeamReadResponse(
-    TeamInfoResponse team,
-    List<GameReadResponse> games
-) {
-
-}
