+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/global/dto/ApiResponse.java
@@ -1,0 +1,12 @@
+package com.spoticket.teamstadium.global.dto;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public record ApiResponse<T>(
+    int status,
+    String msg,
+    T data
+) {
+
+}
