--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/infrastructure/SeatRepository.java
@@ -1,12 +1,0 @@
-package com.spoticket.teamstadium.infrastructure;
-
-
-import com.spoticket.teamstadium.domain.Seat;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SeatRepository extends JpaRepository<Seat, UUID> {
-
-}
