--- com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/infrastructure/TeamRepository.java
@@ -1,12 +1,0 @@
-package com.spoticket.teamstadium.infrastructure;
-
-
-import com.spoticket.teamstadium.domain.Team;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TeamRepository extends JpaRepository<Team, UUID> {
-
-}
