+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/infrastructure/repository/jpa/JpaSeatRepository.java
@@ -1,0 +1,12 @@
+package com.spoticket.teamstadium.infrastructure.repository.jpa;
+
+
+import com.spoticket.teamstadium.domain.model.Seat;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface JpaSeatRepository extends JpaRepository<Seat, UUID> {
+
+}
