+++ com.spoticket.team-stadium/src/main/java/com/spoticket/teamstadium/infrastructure/repository/jpa/JpaStadiumRepository.java
@@ -1,0 +1,12 @@
+package com.spoticket.teamstadium.infrastructure.repository.jpa;
+
+
+import com.spoticket.teamstadium.domain.model.Stadium;
+import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface JpaStadiumRepository extends JpaRepository<Stadium, UUID> {
+
+}
